Compare the grompp warnings against reference
Instead of fixing all warnings the warnings are compared. This
allows to also test mdp settings which produce a warning when
used, but still gurantees that changed grompp behavior is noticed.
Currently the full text of the warning is compared. This can
produce a false positive if the spelling is changed. Between 4.5.5
and 4.6 only one such case occured and as long as it so seldom
it is probably not worth to implement a more complicated
comparison. All one has to do in such a case is remove the
reference.warn and it is automatically regenerated.

All references are produced with 4.5.5

Change-Id: I0e949aede8076caeeccaf83575bcf7f998fe632e