Regenerate freeenergy/relative energy references
Test fail with "tight". These tests also fail withthe original
Gromacs version 920c253 that was used to  generated the reference
values, according to the commit (455484fe) which introduced this test.
Thus it doesn't seem to be a regression. I don't know they were
originally generated to cause "tight" to fail.

Change-Id: I8178eed676a901c1c0a3a7511aec5195c3623cdd