Updated reference values based on new Reference build type
Some of the complex tests failed in recent builds, and closer
investigation showed that it was rather the single precision
reference values deviating, possibly due to more precise
single-precision math functions being implemented as part of
Gromacs-4.6. After checking that this was not related to
errors, all reference values have now been updated based on
a "Reference" Gromacs build type where all optimization has
been turned off.

- There are now log files in each directory describing the host,
compiler, flags, and config (no CPU acceleration) used. The
complex & aminoacids test has been enlarged and modified to
start from a more relaxed conformation.
- The testing script has been modified to use relative tolerances
of 0.001 and absolute ones of 0.05 both for energies and forces,
and there are now both -tight and -relaxed options to move it 10x
in either direction.
- gmxtest.pl now works correctly no matter how many cores there
are on the system it is called for (or how many the user asks for)
by adjusting it internally and rerunning mdrun if necessary.

Change-Id: I83993bd1520abfb2c0cd15f59f6e642ce9e5f474