Update reference .mdp and warning files for 4.6
The new files were copied from the mdout.mdp produced by
60790b19. This will mean the check for matching .mdp
post-processing output is meaningful in future. This
helps catch changes in defaults, etc. that we might want
to react to. Several reference.warn files had to change
accordingly.

If we don't do this, we have to remove the check
for .mdp matching, because we can't provide a regression
test set to users that tells them to check every .mdp
file when there is no value in that.

Change-Id: I8122734ea1bf24f98ff06f1562913343d40c8f90