Free energy tests now pass more stringent grompp error checking
In https://gerrit.gromacs.org/#/c/2015/ a more stringent error
checking is performed on free enery mdp options. Specifically, there is
now a default value of -1 for init-lambda-state, so that it can
be checked whether an initial lambda state (either through init-lambda
or init-lambda-state, but not both) is set.

This broke a few tests in the regression test: those are fixed here.

Change-Id: Ib64292b36a735d306f84500c380044047ae2b566