Fix regressiontests for new default to Verlet scheme
Most tests were OK because their input .mdp files already specified
the old default. pdb2gmx were not, because they did an EM with an
ancient .mdp file that used the default cut-off scheme. This is now
replaced with one that specifies the Group scheme. This will
need re-working later.

Change-Id: Id0efdffcbf0d30b1fc591185468512d126cbd1f2