For simple ns-search run in serial
Because PD gets removed running in serial is often the only
option for ns-type=simple. For verlet OpenMP is automatically
used by mdrun with tMPI. But for velet+MPI and group running in
serial is the only option. Changed mdparams into a sub so that
the number of mpi nodes can be changed for reruns.

Change-Id: I66eec6906a880dd029572c7aad20d63e648d9894