Compare tpr files
The result of gmxcheck for the tpr files wasn't
checked and thus the test always passed.

Also report the check tpr file names, to see the output
in Jenkins.

Fixes those issues where tpr files and mpd files don't agree.
1) calc-lambda-neighbors default is 1. But if it is missing in
a tpr file it is 0 because this corresponds to previous behavior.
2) lmc-seed cannot be -1 (random) for tests
3) Previous rlistlong was decreased automatically if rcoulomb
and rvdw is smaller. This isn't the case anymore
4) init_fep_state differs for old tpr files which still used
init-lambda. Logic seems to produce same result for actually
values used during simulation.

Change-Id: I47d9750be4a0fb994762701a826bb08a8f87c5b9