Compare tpr files
The result of gmxcheck for the tpr files wasn't
checked and thus the test always passed.

Also report the check tpr file names, to see the output
in Jenkins.

Fixes those issuues where tpr files and mpd files don't agree.
1) calc-lambda-neighbors default is 1. But if it is missing in
a tpr file it is 0 because this corresponds to previous behavior.

Change-Id: I47d9750be4a0fb994762701a826bb08a8f87c5b9