Add option to specify mdrun -npme
mdrun -npme triggers a fatal error for an input file that does not use
PME, so the machinery has to prepare for this. Also, mdrun -npme 1
gives a fatal error with only two ranks, so only try -npme when there
are more than two ranks. Also, manage the gpuid string as needed,
now that the number of PP and MPI ranks are not necessarily equal.

Change-Id: Iac0f4ff6705cce1ec156fbb6424c4bf852729816