Fix use of link and rename in regression test harness
When generating new reference data, the script relies on recognizing
that reference file(s) are missing, warning the user, and assuming
that the user intends to generate new reference data. This should be
done with copy semantics. rename is not right because other parts of
the script rely on output files still being there and having the right
name. Using link is apparently not right either, because my helper
script that cleaned once before generating single and then double
reference data has led to the generation of reference_s* files
containing data generated with double precision GROMACS. That makes no
sense, even on a network file system, but having to re-generate these
is apparently my punishment for daring to automate this tedious task.

Naturally, perl does not provide an inbuilt portable copy function,
which is probably why we used link in the first place, so we have to
import File::Copy. I don't like having to do that, but it is not hard
to install the File::Copy module from CPAN, and most distros will have
done that with perl anyway.

Change-Id: I0288e841496aa76efb1a03ef608cce05810e5fc4