Made gmxtest.pl less verbose.
By default gmxtest.pl used to start with a high verbosity
which lead to burying some of the summaries. Now it is less verbose,
defaulr level 2, and users can specify the verbosity at as a
command line argument.

All the summaries are in addition printed at the end for eaiser review.

Change-Id: I5ffbb62f3477adbadda8b20b939b59a6d0e6c5c3