Update naming of things to do with parallelism
We should refer to ranks, rather than threads, processes or
processors, in a way that is similar to how we now do things in the
output of mdrun.

One regexp is now more general, because there's still a string in
mdrun (which should change at some point) that refers to threads. We
don't want to have to update this script when that string changes.

No other functionality changes here, just clean-up.

Change-Id: I67069db5077c9f0d50de7ec0e448fd6c575364de