Updates for grompp-warning checking logic
Removed the use of combining stdout and stderr on shell
command lines, since at least on Windows these can interleave,
and that breaks the assumptions of the subsequent checks for
consistency of the grompp warnings with those of the reference
run. No other examples of combining stdout and stderr remain in
this script.

grompp writes its numbered warnings only to stderr (via the
warning() function), so only stderr needs to be parsed for
comparing those warnings. Updated the implementation of
warnings-comparison accordingly.

Added more verbose output when warnings don't match, which I
needed in order to find out what this problem was.

Hard-coded fewer filenames, and updated some variable names to be
grompp_out, ref_warn etc.

Change-Id: Iac2a9363283f48f0cd76e72f56bfef6a479519fc