Remove the need for twin-range from regressiontests
Removed complex/fe_test because complex/nbnxn-free-energy already
essentially duplicates it, and freeenergy/* also test free-energy
calculations with the group scheme.

Modified all kernel tests to use single-range lists. This means that
the cutoff-based hacks for "None" no longer work, but these are pretty
much only useful for kernel debugging. Detecting e.g. that upcoming
SIMD-support changes are broken is not something that requires us to
be sure that we can isolate a bug to respectively VDW or Coulomb
interactions.

Change-Id: I906875fe1ec564143214d6f25c00b43deba6f85f