Accommodate table-handling transition
We will move handling of user tables to grompp, so here we temporarily
permit the regressiontest harness to discover how it should run the
tests for group-scheme kernels that require short-ranged tables.
Integration-style tests for bonded-interaction tables will be added to
the source repository, so only checks for handling short-ranged tables
need to happen in regressiontests repository.

Refs #1347

Change-Id: I96f2b7ea792fea983fc60ff6b9acf3d0edeca8e9