[WIP] Regression tests for essential dynamics
These tests check whether the basic functionality of the
essential dynamics module works as expected. There is one
test for each of the six flavours of essential dynamics:
- fixed-step linear expansion (LINFIX)
- acceptance linear expansion (LINACC)
- fixed-step radius expansion (RADFIX)
- acceptance radius expansion (RADACC)
- acceptance radius contraction (RADCON)
- flooding (FLOODING)
The first five tests use a simple n-butane molecule in vacuum,
very similar to the system used in (1), while the last test uses
a more realistic scenario of a guanylin molecule solvated in water.

(1) OF Lange, LV Schafer, H Grubmuller,
    J. Comput. Chem. 27: 1693-1702 (2006)

Change-Id: Iab7a0e72f836fb22b6e956cdc60d204089079d45