Made bonded-interaction testing consistent
Tests of normal dihedrals have been inactive forever
because -DDIH doesn't work with #ifdef DIH. Made
that work, regenerating those test cases.

Changed DIH and RYCKAERT to be always plural, so that all
the kinds of interaction work the same way.

Eliminated use of MOL1 so that all test cases specify the number of
molecules the same way (ie hard-coded in topol.top).

Change-Id: Icc50b962e5a7e4d16ed2ae8e723862eb8c085f1b