Permit gmxtest.pl to test PME on GPUs
Introduce some temporary functionality to check whether mdrun is
capable of running PME on GPUs, which will avoid needing to do a lot
of Jenkins cross-verification (and humans triggering them). (We can
remove the check at a later time if we think we need to.) If the code
can run PME on GPUs, that can be tested in the same way as the GPU
code.

Following the example of the way we re-run tests so that we also test
CPU non-bonded kernels on hardware with GPUs, we arrange to test PME
on both CPU and GPU when testing of GPUs is required. A better harness
could perhaps detect whether GPUs are present (or were used by default
for PME), but the current implementation works for Jenkins, at least.

The continuation tests such as tip4p_continue now support re-running.

Accordingly, extracted reused functionality into
prepare_run_with_different_task_decomposition.

Refs #2124, #1587

Change-Id: I64c45c62f96d06f057a6f886727fc5852b84b9ba