Restrict gmxtest.pl from testing PME on GPUs
Introduce functionality to check whether mdrun is capable of running
PME on GPUs, which will avoid needing to do a lot of Jenkins
cross-verification (and humans triggering them).
(We can remove the check at a later time if we think we need to.)

For now, if the mdrun is capable of running PME on GPUs,
then "-pme cpu" is appended to its command line, forcing it
to run PME on CPU, therefore preserving the current testing scope
and ignoring the PME GPU codepath. The reasoning is that new
test functionality (such as PME on GPU overall tests) would be better
implemented using the GTest framework, allowing for clearer
scopes of separate test cases, rather than adding more complicated
"re-running" logic within gmxtest.pl.

Refs #2124, #1587

Change-Id: Ie6f327d139b09b61eeb969655c18d9750c2bdb1d