Annotate test cases that cannot run on GPUs
Jenkins runs gmxtest.pl in post-submit testing by specifying values
for -gpu_id for it to use. mdrun now interprets -gpu_id as requiring a
GPU run, and gives a fatal error if a group-scheme tpr is supplied.

The group-scheme test cases are now updated with an annotation (like
max-mpi-ranks) in the cases where GPUs are not supported (ie group
scheme), so the runner can know to omit specifying an erroneous
-gpu_id. Once everything runs with the Verlet scheme, then these
annotations will no longer be required, and will disappear naturally.

The feature where test cases that did run on a GPU are then run again
with -nb cpu is modified so that those repeat runs do not specify an
erroneous -gpu_id.

Refs #2067

Change-Id: Ia16a59f57be0770be921b072b8f51639a9cc6dc0