Fix gmxtest.pl handling of GPU IDs
For example, mdrun -ntmpi 2 -gpu_id 2 would not start 2 PP ranks on
GPU ID 2, which was what we'd intended it to do. Instead, the automatic
repeater would make it run with one rank.

mdrun -ntmpi 2 -gpu_id 01 was OK though.

This change would permit a Jenkins matrix to specify any of

gcc-4.8 gpu gpu_id=1
gcc-4.8 gpu nranks=2 gpu_id=012
gcc-4.8 gpu npme=1 nranks=3 gpu_id=0

and ensures that we get the number and flavour of ranks specified.

This should mean it is no longer desirable to rerun a test case
when the number of ranks and the GPU ID choice do not work.

Refs #2134

Change-Id: Idba4a7de2069a67adb4cdf8b3e025a35d13c1f14