Added tests of four more modern force fields.
The pdb2gmx test in the regression tests would only test
the older force fields gromos43a1, oplsaa and gromos53a6.
Added testing of amber99sb-ildn, amber03, charmm27, and
gromos54a7. Unfortunately the order of terms in the energy
files used for checking is now changed and therefore it seems
there are more changes in the energies than there are in fact.

Change-Id: I3440f5825958c0f0575860e1564a697cce266faf