Update handling of GPU runs
Renamed no-gpu-support to no-nb-gpu-support now that PME is supported
on GPUs.

Updated logic that determines whether to re-run a PME test so that we
test both code paths. The old code had assumptions that were no longer
applicable. Annoted the PME test cases that are known to be able to
run on a GPU, and added logic to check whether in fact those test
cases did run on a GPU, before rerunning them.

Because PME on GPUs requires NB on GPUs, eliminated one of the rerun
paths.

Change-Id: I7c26458bf907e6128bf86b2c5da0b90b6d200a8c