Fix make check
Apparently this hasn't worked for over three years, because
the behaviour of eval is not to set the return value in $?. Now the
map of the instructions to execute correctly accumulates whether any
of the fail.

Also avoided use of undefined variables when the reference mdp file
has more lines than the mdp file output by grompp, which will happen
when features like implicit solvation are removed.

Change-Id: If2e015a50900b92cb5da42b5d34822d52423293c