Half number of steps for position-restraints
It is the only complex or simple test which fails force check
with tight with 8 cores, suggesting it is more unstable than
most tests. With GMX_OCL_NB_CLUSTER_SIZE=4 test failed.
It was fine with relaxed set to 2. And is now fine even with
tight. Suggesting that this was caused by unstability of the
test not a problem in the code.

Related #2030

Change-Id: Iebf46b085aa4fea36ac9eddc4ac3fa72846f88f7