Updated normal modes output for sw-dimer.
Due to a bug in mtop_util.cpp only the first molecule in the
sw-dimer was used in a normal mode analysis. This is fixed in patch
https://gerrit.gromacs.org/#/c/8631/.

Fixed a bug in gmxtest.pl when comparing two xvg files where the
reference value is zero (it computed the relative difference and
divided by zero). This only showed up in the new output file
normal_modes/sw_dimer/eigenfreq.xvg.

Part of #2720

Change-Id: Iae990bbff2c19c28310618f2cadf1157a62242d0