Add ability for regressiontest to detect new DD requirements
With the addition of the GPU halo exchange code, it can happen that the
regressiontest script will fail because a test does not use PME and also
does not work with the required ranks and GPU tasks.

This adds a detection function for this kind of error condition to the
script, and the ability to automatically adjust the number of ranks to
match the number of available GPUs in those cases.

Also added some missing docs for existing functions.

Change-Id: I0c90a3cc875f748b42a33a6e534323bcffffc6b5