Add second try for automatic mpi rank selection
If the automatic setting of mpi ranks fails also for the case of 8
ranks, try another time with 2 ranks instead to get around domain
decomposition limitations.

Also ensure that we have at least two ranks in case of separate PME
ranks, to ensure that test cases that are supposed to use domain
decomposition actually use it.

Change-Id: Ia6d1da6c89cc57c204729d303fc2b3fc7c4d359f